half4 replaced float4 for Frags. Slight improvement.

+++ Assignment 1/Assignment_1/Assets/Objects/Standard Assets/Characters/ThirdPersonCharacter/Materials/EthanShader.shader
@@ -94,7 +94,7 @@ Shader "EthanShader"
 			}
 
 			// normal mapping pixel shader
-			float4 PS_NormalMapping(VSOutput a_Input) : COLOR
+			half4 PS_NormalMapping(VSOutput a_Input) : COLOR
 			{
 				// calculate vector to camera
 				float3 toCamera = normalize(_vecCameraPos.xyz - a_Input.posWorld.xyz);
@@ -216,7 +216,7 @@ Shader "EthanShader"
 			}
 
 			// normal mapping pixel shader
-			float4 PS_NormalMapping(VSOutput a_Input) : COLOR
+			half4 PS_NormalMapping(VSOutput a_Input) : COLOR
 			{
 				// calculate vector to camera
 				float3 toCamera = normalize(_vecCameraPos.xyz - a_Input.posWorld.xyz);
